CI: Add MATRIX_BUILD_LIBPCAP to pass on FreeBSD.
The previously implemented approach was to let ./configure detect
libpcap, which would be either a local libpcap (if the caller has
pre-built it in ../libpcap) or a system libpcap (if it has not). In
particular, the Buildbot master setup arranged a local libpcap, which
defaulted to GCC [1]. The Buildbot workers that could compile tcpslice
varied in which libpcap they were able to use:

* freebsd-aarch64: system only [1]
* linux-aarch64: both
* linux-armv7l: both
* linux-ppc64le: both
* netbsd-aarch64: local only
* netbsd-amd64: local only

To simplify the builder commands and to allow customisation at the same
time, copy the MATRIX_BUILD_LIBPCAP logic from tcpdump and use it in
worker-specific environment variables as necessary.

A useful side effect of this change is compiling the local libpcap using
the same CC as for the current round of tcpslice. This works around a
linking issue [1], restores tcpslice-freebsd-aarch64 and allows it and
other builders to test two additional permutations.

1: FreeBSD shortly before release 13.0 started to have an odd issue, in
that tcpslice compiled using Clang failed to link with local libpcap
compiled using GCC:

ld: error: ./pcap.c:736:(.text+0x116C): unknown relocation (313) against
symbol pcap_netmap_findalldevs

ld: error: ./pcap.c:2335:(.text+0x1548): unknown relocation (313)
against symbol pcap_netmap_create

ld: error: ./pcap.c:3502:(.text+0x2210): unknown relocation (313)
against symbol __stderrp

(and so on)

+++ build_matrix.sh
@@ -44,20 +44,32 @@ for CC in ${MATRIX_CC:-gcc clang}; do
         echo '(skipped)'
         continue
     fi
-    COUNT=$((COUNT+1))
-    echo_magenta "===== SETUP $COUNT: CC=$CC ====="
-    # LABEL is needed to build the travis fold labels
-    LABEL="$CC"
-    # Run one build with the setup environment variable: CC
-    ./build.sh
-    echo 'Cleaning...'
-    travis_fold start cleaning
-    make distclean
-    rm -rf "$PREFIX"/*
-    git status -suall
-    # Cancel changes in configure
-    git checkout configure
-    travis_fold end cleaning
+    for BUILD_LIBPCAP in ${MATRIX_BUILD_LIBPCAP:-yes no}; do
+        COUNT=$((COUNT+1))
+        echo_magenta "===== SETUP $COUNT: CC=$CC BUILD_LIBPCAP=$BUILD_LIBPCAP ====="
+        if [ "$BUILD_LIBPCAP" = yes ]; then
+            echo_magenta "Build libpcap (CMAKE=no)"
+            (cd ../libpcap && CMAKE=no ./build.sh)
+        else
+            echo_magenta 'Use system libpcap'
+            rm -rf "$PREFIX"/*
+            if [ -d ../libpcap ]; then
+                make -C ../libpcap distclean || :
+            fi
+        fi
+        # LABEL is needed to build the travis fold labels
+        LABEL="$CC.$BUILD_LIBPCAP"
+        # Run one build with the setup environment variable: CC
+        ./build.sh
+        echo 'Cleaning...'
+        travis_fold start cleaning
+        make distclean
+        rm -rf "$PREFIX"/*
+        git status -suall
+        # Cancel changes in configure
+        git checkout configure
+        travis_fold end cleaning
+    done
 done
 rm -rf "$PREFIX"
 echo_magenta "Tested setup count: $COUNT"
